Rename lockfilescalibr package
