+++ internal/scalibrextract/language/java/pomxmlnet/extractor.go
