+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/empty.xml
