+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/interpolation.xml
