+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/invalid-syntax.xml
