+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/not-pom.txt
