+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/one-package.xml
