+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/parent/pom.xml
