+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/transitive.xml
