+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/two-packages.xml
