+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/with-dependency-management.xml
