+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/with-parent.xml
