+++ internal/scalibrextract/language/java/pomxmlnet/testdata/maven/with-scope.xml
