+++ internal/scalibrextract/language/java/pomxmlnet/testdata/universe/basic-universe.yaml
