+++ internal/scalibrextract/language/javascript/nodemodules/extractor.go
