+++ internal/scalibrextract/language/osv/osvscannerjson/extractor.go
