+++ internal/scalibrextract/language/osv/osvscannerjson/extractor_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/google/osv-scalibr/extractor"
 	"github.com/google/osv-scalibr/testing/extracttest"
-	"github.com/google/osv-scanner/internal/lockfilescalibr/language/osv/osvscannerjson"
+	"github.com/google/osv-scanner/internal/scalibrextract/language/osv/osvscannerjson"
 	"github.com/google/osv-scanner/pkg/models"
 )
