+++ internal/scalibrextract/language/osv/osvscannerjson/metadata.go
