+++ internal/scalibrextract/language/osv/osvscannerjson/testdata/empty.json
