+++ internal/scalibrextract/language/osv/osvscannerjson/testdata/multiple-packages-with-vulns.json
