+++ internal/scalibrextract/language/osv/osvscannerjson/testdata/one-package-commit.json
