+++ internal/scalibrextract/language/osv/osvscannerjson/testdata/one-package.json
