+++ internal/scalibrextract/vcs/gitrepo/extractor_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/google/osv-scalibr/extractor"
 	"github.com/google/osv-scalibr/testing/extracttest"
-	"github.com/google/osv-scanner/internal/lockfilescalibr/vcs/gitrepo"
+	"github.com/google/osv-scanner/internal/scalibrextract/vcs/gitrepo"
 )
 
 func TestExtractor_Extract(t *testing.T) {
