+++ internal/scalibrextract/vcs/gitrepo/testdata/example-git/a.txt
