+++ internal/scalibrextract/vcs/gitrepo/testdata/example-git/git-hidden/info/exclude
