+++ internal/scalibrextract/vcs/gitrepo/testdata/example-git/git-hidden/objects/16/b14f5da9e2fcd6f3f38cc9e584cef2f3c90ebe
