+++ internal/scalibrextract/vcs/gitrepo/testdata/example-git/git-hidden/objects/86/2ac4bd2703b622e85f29f55a2fd8cd6caf8182
