+++ internal/scalibrextract/vcs/gitrepo/testdata/example-git/git-hidden/objects/bf/8fbfe5a434c007b640c12d920683cb19a7b2b9
