+++ internal/scalibrextract/vcs/gitrepo/testdata/example-git/git-hidden/refs/heads/main
