+++ internal/scalibrextract/vcs/gitrepo/testdata/example-not-git/a.txt
