+++ internal/scalibrextract/vcs/gitrepo/testdata/example-not-git/git-hidden/b.txt
